Rename akari_controller to akari_client
