+++ samples/akari_sample/script/2b_gpio_input.py
@@ -9,7 +9,7 @@ Created on 2021/06/11
 
 import time
 
-from akari_controller import AkariClient
+from akari_client import AkariClient
 
 
 def main() -> None:
