+++ samples/akari_sample/script/4a_display_color.py
@@ -11,8 +11,8 @@ Created on 2021/06/11
 import time
 import random
 
-from akari_controller import AkariClient
-from akari_controller.color import Colors, Color
+from akari_client import AkariClient
+from akari_client.color import Colors, Color
 
 
 def main() -> None:
