+++ sdk/akari_client/tests/akari_controller_tests/test_m5serial_server_py.py
@@ -1,8 +1,8 @@
 from typing import cast
 
 import pytest
-from akari_controller.m5serial_communicator import M5SerialCommunicator
-from akari_controller.m5stack_client import M5StackSerialClient, _PinOut
+from akari_client.m5serial_communicator import M5SerialCommunicator
+from akari_client.m5stack_client import M5StackSerialClient, _PinOut
 
 from .mocks import MockM5Communicator
